糅合了项目详情页面。提测工单页面和上线工单页面。同时保留提测工单和上线工单单独的页面

+++ src/main/java/com/facebank/usersupport/online/process/service/ILaunchFormService.java
@@ -15,7 +15,7 @@ public interface ILaunchFormService {
 
     LaunchFormModel selectByFormId(Long formId);
 
-    int insertLaunchForm(LaunchFormModel launchFormModel);
+    Long insertLaunchForm(LaunchFormModel launchFormModel);
 
     int updateLaunchForm(LaunchFormModel launchFormModel);
 }
